m88: Fix build of skia-safe
