+++ skia-org/src/drivers/gl.rs
@@ -4,7 +4,7 @@ use skia_safe::{gpu, Budgeted, Canvas, ImageInfo, Surface};
 use std::path::Path;
 
 pub struct OpenGL {
-    context: gpu::Context,
+    context: gpu::DirectContext,
 }
 
 impl DrawingDriver for OpenGL {
@@ -12,7 +12,7 @@ impl DrawingDriver for OpenGL {
 
     fn new() -> Self {
         Self {
-            context: gpu::Context::new_gl(None).unwrap(),
+            context: gpu::DirectContext::new_gl(None, None).unwrap(),
         }
     }
