+++ skia-safe/src/gpu.rs
@@ -10,14 +10,14 @@ pub use self::backend_surface_mutable_state::*;
 pub mod context_options;
 pub use self::context_options::ContextOptions;
 
-mod context;
-pub use self::context::*;
-
 #[cfg(feature = "d3d")]
 pub mod d3d;
 
 mod direct_context;
 pub use self::direct_context::*;
+
+#[deprecated(note = "Use DirectContext", since = "0.0.0")]
+pub type Context = DirectContext;
 
 mod driver_bug_workarounds;
 pub use self::driver_bug_workarounds::DriverBugWorkarounds;
